apps/property-api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
notion-client>=2.2.0
